+++ src/fr/maxlego08/zshop/api/buttons/ItemConfirmButton.java
@@ -14,8 +14,6 @@ public interface ItemConfirmButton extends EconomyAction {
 
     boolean enableLog();
 
-    List<String> getConfirmMessages();
-
     String getName();
 
     String getInventoryConfirm();