+++ src/fr/maxlego08/zshop/buttons/ZItemConfirmButton.java
@@ -1,5 +1,6 @@
 package fr.maxlego08.zshop.buttons;
 
+import fr.maxlego08.menu.api.requirement.Action;
 import fr.maxlego08.menu.api.utils.Placeholders;
 import fr.maxlego08.menu.button.ZButton;
 import fr.maxlego08.menu.inventory.inventories.InventoryDefault;
@@ -10,10 +11,8 @@ import fr.maxlego08.zshop.api.economy.ShopEconomy;
 import fr.maxlego08.zshop.api.history.History;
 import fr.maxlego08.zshop.api.history.HistoryType;
 import fr.maxlego08.zshop.history.ZHistory;
-import fr.maxlego08.zshop.placeholder.Placeholder;
 import fr.maxlego08.zshop.save.LogConfig;
 import fr.maxlego08.zshop.zcore.logger.Logger;
-import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
 
@@ -24,21 +23,21 @@ public class ZItemConfirmButton extends ZButton implements ItemConfirmButton {
     private final ShopPlugin plugin;
     private final ShopEconomy shopEconomy;
     private final double price;
-    private final List<String> commands;
     private final boolean enableLog;
-    private final List<String> messages;
     private final String name;
     private final String inventoryConfirm;
+    private final String withdrawReason;
+    private final List<Action> confirmActions;
 
-    public ZItemConfirmButton(ShopPlugin plugin, ShopEconomy economy, double price, List<String> commands, boolean enableLog, List<String> messages, String name, String inventoryConfirm) {
+    public ZItemConfirmButton(ShopPlugin plugin, ShopEconomy economy, double price, boolean enableLog, String name, String inventoryConfirm, String withdrawReason, List<Action> confirmActions) {
         this.plugin = plugin;
         this.shopEconomy = economy;
         this.price = price;
-        this.commands = commands;
         this.enableLog = enableLog;
-        this.messages = messages;
         this.name = name;
         this.inventoryConfirm = inventoryConfirm;
+        this.withdrawReason = withdrawReason;
+        this.confirmActions = confirmActions;
     }
 
     @Override
@@ -57,11 +56,6 @@ public class ZItemConfirmButton extends ZButton implements ItemConfirmButton {
     }
 
     @Override
-    public List<String> getConfirmMessages() {
-        return this.messages;
-    }
-
-    @Override
     public void buy(Player player, int amount) {
 
         ZShopManager manager = (ZShopManager) this.plugin.getShopManager();
@@ -73,15 +67,16 @@ public class ZItemConfirmButton extends ZButton implements ItemConfirmButton {
             return;
         }
 
-        this.shopEconomy.withdrawMoney(player, currentPrice);
+        this.shopEconomy.withdrawMoney(player, currentPrice, this.withdrawReason);
 
         /* END ECONOMY CHECK */
 
-        this.getMessages().forEach(message -> manager.message(this.plugin, player, message.replace("%player%", player.getName())));
-        this.getCommands().forEach(command -> {
-            command = command.replace("%player%", player.getName());
-            Bukkit.dispatchCommand(Bukkit.getConsoleSender(), Placeholder.getPlaceholder().setPlaceholders(player, command));
-        });
+        Placeholders placeholders = new Placeholders();
+        placeholders.register("player", player.getName());
+
+        for (Action confirmAction : this.confirmActions) {
+            confirmAction.preExecute(player, this, this.plugin.getIManager().getFakeInventory(), placeholders);
+        }
 
         if (LogConfig.enableLog || this.enableLog) {
 
@@ -110,11 +105,6 @@ public class ZItemConfirmButton extends ZButton implements ItemConfirmButton {
     public void onClick(Player player, InventoryClickEvent event, InventoryDefault inventory, int slot, Placeholders placeholders) {
         super.onClick(player, event, inventory, slot, placeholders);
         this.plugin.getShopManager().openConfirm(player, this, this.inventoryConfirm);
-    }
-
-    @Override
-    public List<String> getMessages() {
-        return messages;
     }
 
     @Override